re/lostIt/requirements.txt
uuid
sqlalchemy
sqlalchemy_utils
pycrypto
